+++ Rdmp.Core.Tests/CohortCreation/CohortCompilerTests.cs
@@ -11,9 +11,12 @@ using NUnit.Framework;
 using Rdmp.Core.CohortCreation;
 using Rdmp.Core.CohortCreation.Execution;
 using Rdmp.Core.CohortCreation.Execution.Joinables;
+using Rdmp.Core.CommandExecution;
+using Rdmp.Core.CommandExecution.AtomicCommands;
 using Rdmp.Core.Curation.Data;
 using Rdmp.Core.Curation.Data.Aggregation;
 using Rdmp.Core.Curation.Data.Cohort.Joinables;
+using Rdmp.Core.DataExport.Data;
 
 namespace Rdmp.Core.Tests.CohortCreation;
 
@@ -22,7 +25,7 @@ public class CohortCompilerTests : CohortIdentificationTests
     [Test]
     public void AddSameTaskTwice_StaysAtOne()
     {
-        var compiler = new CohortCompiler(cohortIdentificationConfiguration);
+        var compiler = new CohortCompiler(new ThrowImmediatelyActivator(RepositoryLocator,null), cohortIdentificationConfiguration);
         container1.AddChild(aggregate1, 0);
         try
         {
@@ -66,7 +69,7 @@ public class CohortCompilerTests : CohortIdentificationTests
     [Test]
     public void AddContainer_StaysAtOne()
     {
-        var compiler = new CohortCompiler(cohortIdentificationConfiguration);
+        var compiler = new CohortCompiler(new ThrowImmediatelyActivator(RepositoryLocator, null), cohortIdentificationConfiguration);
         rootcontainer.AddChild(aggregate1, 1);
 
         compiler.AddTask(rootcontainer, null); //add the root container
@@ -144,7 +147,7 @@ public class CohortCompilerTests : CohortIdentificationTests
 
             //Joinable:aggregate5 (patient index table, the other Aggregates could JOIN to this)
 
-            var compiler = new CohortCompiler(cohortIdentificationConfiguration);
+            var compiler = new CohortCompiler(new ThrowImmediatelyActivator(RepositoryLocator, null), cohortIdentificationConfiguration);
             rootcontainer.AddChild(aggregate1, 1);
             rootcontainer.AddChild(container1);
             container1.Order = 2;
@@ -196,5 +199,78 @@ public class CohortCompilerTests : CohortIdentificationTests
             joinable.DeleteInDatabase();
             aggregate5.DeleteInDatabase();
         }
+    }
+
+    [Test]
+    public void TestCompilerWithInternalCatalogues()
+    {
+        var compiler = new CohortCompiler(new ThrowImmediatelyActivator(RepositoryLocator, null), cohortIdentificationConfiguration);
+        testData.catalogue.IsInternalDataset = true;
+        testData.catalogue.SaveToDatabase();
+
+        rootcontainer.AddChild(aggregate1, 1);
+        rootcontainer.AddChild(container1);
+        container1.Order = 2;
+        container1.SaveToDatabase();
+
+        cohortIdentificationConfiguration.RootCohortAggregateContainer_ID = rootcontainer.ID;
+        cohortIdentificationConfiguration.SaveToDatabase();
+
+        compiler.AddAllTasks(true);
+        Assert.That(compiler.Tasks.Keys.Select(k => k.State).Where(s => s == CompilationState.Crashed).ToList(), Has.Count.EqualTo(3));
+        Assert.That(compiler.Tasks.Keys.Where(s => s.State == CompilationState.Crashed).Select(k => k.CrashMessage.Message).ToList().Any(m => m.Contains(" is marked as Internal. Internal Catalogues cannot be used in Cohort Identification Configurations.")));
+    }
+
+
+    [Test]
+    public void TestCompilerWithProjectSpecificCatalogues_Fail()
+    {
+        var compiler = new CohortCompiler(new ThrowImmediatelyActivator(RepositoryLocator, null), cohortIdentificationConfiguration);
+        
+        var proj = new Project(DataExportRepository, "TestCompilerWithProjectSpecificCatalogues_Fail");
+        proj.SaveToDatabase();
+
+        //having an issue making it project specific
+        var cmd = new ExecuteCommandMakeCatalogueProjectSpecific(new ThrowImmediatelyActivator(RepositoryLocator), aggregate1.Catalogue, proj, true);
+        Assert.DoesNotThrow(() => cmd.Execute());
+        rootcontainer.AddChild(aggregate1, 1);
+        rootcontainer.AddChild(container1);
+        container1.Order = 2;
+        container1.SaveToDatabase();
+
+        cohortIdentificationConfiguration.RootCohortAggregateContainer_ID = rootcontainer.ID;
+        cohortIdentificationConfiguration.SaveToDatabase();
+
+        compiler.AddAllTasks(true);
+        Assert.That(compiler.Tasks.Keys.Select(k => k.State).Where(s => s == CompilationState.Crashed).ToList(), Has.Count.EqualTo(3));
+        Assert.That(compiler.Tasks.Keys.Where(s => s.State == CompilationState.Crashed).Select(k => k.CrashMessage.Message).ToList().Contains("Catalogue BulkData is marked as Project Specific, but this Cohort Identification Configurations is not associated with the same Project."));
+    }
+    [Test]
+    public void TestCompilerWithProjectSpecificCatalogues_Pass()
+    {
+        var compiler = new CohortCompiler(new ThrowImmediatelyActivator(RepositoryLocator, null), cohortIdentificationConfiguration);
+
+        var proj = new Project(DataExportRepository, "TestCompilerWithProjectSpecificCatalogues_Fail");
+        proj.SaveToDatabase();
+
+        //having an issue making it project specific
+        var cmd = new ExecuteCommandMakeCatalogueProjectSpecific(new ThrowImmediatelyActivator(RepositoryLocator), aggregate1.Catalogue, proj, true);
+        Assert.DoesNotThrow(() => cmd.Execute());
+        rootcontainer.AddChild(aggregate1, 1);
+        rootcontainer.AddChild(container1);
+        container1.Order = 2;
+        container1.SaveToDatabase();
+
+        cohortIdentificationConfiguration.RootCohortAggregateContainer_ID = rootcontainer.ID;
+        cohortIdentificationConfiguration.SaveToDatabase();
+
+        var associateCmd = new ExecuteCommandAssociateCohortIdentificationConfigurationWithProject(new ThrowImmediatelyActivator(RepositoryLocator));
+        associateCmd.SetTarget(proj);
+        associateCmd.SetTarget(cohortIdentificationConfiguration);
+        Assert.DoesNotThrow(() => associateCmd.Execute());
+
+        compiler.AddAllTasks(true);
+        Assert.That(compiler.Tasks.Keys.Select(k => k.State).Where(s => s == CompilationState.Crashed).ToList(), Has.Count.EqualTo(2));
+        Assert.That(!compiler.Tasks.Keys.Where(s => s.State == CompilationState.Crashed).Select(k => k.CrashMessage.Message).ToList().Contains("Catalogue BulkData is marked as Project Specific, but this Cohort Identification Configurations is not associated with the same Project."));
     }
 }